backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
courlan==1.3.2
cryptography==46.0.6
dateparser==1.4.0
distro==1.9.0
feedparser==6.0.12
google-auth==2.49.1
google-genai==1.68.0
googlenewsdecoder==0.1.7
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
justext==3.0.2
lxml==6.0.2
lxml-html-clean==0.4.4
pillow==12.1.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
regex==2026.2.28
requests==2.33.0
selectolax==0.4.7
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tld==0.13.2
trafilatura==2.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0